Updated HUD functionality
Added a custom Game Instance, a custom Player Controller, and created a local subsystem to act as the UI manager. This subsystem now has a function to create the HUD, which is called by the Player Controller. The inventory is now an overlay added to the main HUD, which is the only canvas, and the UI Manager subsystem handles the entire process of managing the UIs.